外部改动提示条队列改存页 id，裁决不再落到别页上
问题：提示条是非模态的。队列 [A,C] 挂着时关掉 A，旧实现存的裸下标 [0,2] 就此
左移——「0」现在指 B：条子改名成 B（它从未被外部改过、也不该被裁决），〔忽略〕
把 B 的比对戳推到磁盘现状。此后 B 再被外部改动时，P63 保存守卫看到
「磁盘 == 记录戳」直接落盘覆盖，一次提示都不弹；镜像症状是〔重新加载〕把 B 的
未保存正文掏空。与上一条批量关闭（P213）、重命名（P214）同一病根。

原因：external_change: Option<Vec<usize>>，四个生产者记 idx（派发时刻的位置），
三个消费者按下标现取；close_tab_now 只平移 close_tab_confirm 一个字段。

改法：队列与两条裁决消息一律带 TabId，消费者按 id 解析当前位置——解析不到
（页已关）就是彻底的空操作，只把那个陈旧队首摘掉。生产者四处的 idx 换成
tab.id / 消息自带的 tab_id；视图侧「条子展示哪一页」同步按 id 查。

验证：全量 775 绿；clippy 与 cargo doc 均 exit=0。新增
external_change_prompt_ignores_survive_a_queued_tab_closing：三页、A 与 C 入队、
关掉 A、按视图会发的队首值裁决，断言 B 的戳一点没动。反证做了一次**修正**：
第一版把这条测成恒绿——B 的磁盘内容与记录戳本来一致，「被误推戳」根本不可观测；
改成「巡检之后再外部改写 B」并补一条夹具自检（显式对 B 裁决必须看到戳变化），
再把生产者+消费者都退回旧形状，才在行为断言上拿到真红
（left Some((t,33)) vs right Some((t,5))）。既有 5 条外部改动用例的断言由
「等于下标」改为经 prompt_ids 换算的页 id。

+++ crates/app/src/main.rs
@@ -591,10 +591,10 @@ enum Message {
     /// 窗口重新获得焦点：巡检各命名页的 (mtime, size) 戳，外部已改的
     /// 干净活动页静默重载，其余弹提示条由用户裁决
     WindowFocused,
-    /// 提示条「重新加载」：放弃该页未保存内容并从磁盘重读
-    ConfirmExternalReload(usize),
-    /// 提示条「忽略」：以当前磁盘状态重记时间戳，直到下次再变不再提示
-    IgnoreExternalChange(usize),
+    /// 提示条「重新加载」：放弃该页未保存内容并从磁盘重读（P220：携带页 id）
+    ConfirmExternalReload(u64),
+    /// 提示条「忽略」：以当前磁盘状态重记时间戳，直到下次再变不再提示（页 id）
+    IgnoreExternalChange(u64),
     /// 提示条「全部忽略」（聚合态）：队列内所有页一律重记时间戳并收条
     IgnoreAllExternalChanges,
 }
